Issue #3056837 by bramtenhove: Statically cache rendered profile view mode output
In a ton of places we load the profile entity storage and load the profile for a given user.
After that we load the view builder and often render the same output multiple times in a single page request.

I created a service, not sure if it's in the right place, that statically cache the loaded profile and the output
per view mode.

It could be useful to split up the profile loading in a different function and make that static as well. Not part
of this scope though.