TB-1894 - Add Social Magic Login module (#1382)
Adds a service with interface (so others can implement the service as well) that allows developers to request a one-time login links. The logic for this login link is copied from Drupal 8's password reset functionality.

It encodes the destination in base64 because the ?destination parameter didn't work as expected (causing the login controller to be bypassed). Base64 encoding makes the URL parts safe for transmission in URLs as a single argument.

Typehints in the new service are used because it is code that will only run on PHP 7.2+ and has no existing dependants or dependencies.