Issue #3072888 by Kingdutch, xinyuma, jaapjan: visually separate sorting and filtering options.
We add the visual separation of sorting and filtering to show users that
these are different options. The filter title is added in CSS because
it's not easy to reliably change the form structure.

This commit also undoes the changes to the block template because it
breaks the sorting form behaviour (the field would be rendered outside
of the form).