Issue #3075818 by Kingdutch: Exports don't support applications running on multiple servers
This changes the way the file path for exports are determined. By only
storing the path relative to the temporary directory it's possible for
the location of the temporary directory to change. This can happen when
multiple web-servers handle the same export and use a file mount of a
shared file system. In this case the mounting point could be on a
different path containing the export file. Storing the absolute path at
the start of the export would break in this case.

This commit also changes the visibility of the path functions because
these are implementation details of the export and should not be queried
outside of the export.