TB-2352 Convert Full Name search protection to Search API Processor
Instead of using a complex query alter hook we implement a proper search
processor. This will create a field with the value that users without
the permission to view protected names are allowed to see. This allows
us to change the query to simply ignore protected fields. This should
resolve an existing bug where the full name for a user could be guessed.