TB-2352 Add configuration override for search processors
Although a previous commit added the `all_fields` attribute to our
search indexes' most important processors (which we should keep), the
fields are only updated on a index save. This means that the fields will
be removed during a features revert and may then not be added back after
until someone manually edits the search index. To get around this we
require a configuration override.

This override can be removed once we get rid of features.