TB-2352 Add configuration change event subscriber
Changing configuration does not lead to a re-index by default. However,
our configuration affects what values are stored in the search index.
This requires a re-index to be initiated when the configuration changes.
To achieve this an event subscriber is used that routes the change event
for our specific configuration value to mark the search indexes that
have profile entities as stale.