TB-2352 Mark stale in tracker instead of re-index
We have indexes (such as social_all) that have a lot more data than just
user profiles. Thus instead of marking the entire index to be re-indexed
we access the tracker and tell it to mark all items based on a profile
entity as stale when the configuration changes. This ensures we do the
bare minimum needed.