TB-2352 Re-save search indices during profile privacy module installation
The indices are also resaved during module installation to ensure the
processor is properly added. The index is no longer disabled and enabled
because this is slower and not needed. It worked around a validation of
whether values changed but this was in the Search Settings form and not
the Index entity save logic, so re-saving will trigger all we need.