Remove features from Social Tagging module
This removes the features file which ensures it's no longer considered
to be a feature by the feature module. It also adds a default settings
file which will now not be reverted and removes these settings from the
install hook.

Additionally we use a new helper method in the social_search module to
optimise some of our code.