Add file that can store legacy translations
Sometimes the latest version of Open Social changes strings. These
strings will be picked up in the translation workflow. However, the old
strings that were changed or removed will be removed from translation
source files. This can cause translations to go missing for platforms
running older version.

To fix this we create this dummy file that can hold the old versions of
these strings so that they'll remain translatable in Weblate.