DS-6965 Change node_edit_link to render array
The preprocess hook does not need to check what page it's on because
it's not called for the edit or delete page.

This commit also changes the way the URL is rendered to use a render
array. This causes less operations in the twig template (if statement,
string append etc.) and allows the node edit URL to do its own access
checking instead of doing this in a preprocess hook.