Remove usage of SocialLanguageMetadataBubblingUrlGenerator
This service overwrites too many languages in links which causes issues
such as language switchers or links between content translations not
working.

It's removed in favour of a more targetted approach on the Entity level
of entities that appear in the stream and are not translatable (will be
in a future commit).