Overwrite NodeViewBuilder to fix language links
Remove the language argument from all links. It was added in
Issue #2149649 to solve the case where multiple translations in a single
view would link to the incorrect entity. However, this causes the issue
in Open Social that links change the interface language. Removing this
language argument allows the interface translation to remain the same
even when navigating to an entity that only exists in another language.

Perhaps with the move in Drupal core to toUrl this language attribute
can be removed again. It would require a change to
`template_preprocess_links` to try and fallback to URL language for the
hreflang attribute if the link language is not present. An issue could
be opened for this as there are Drupal core tests to catch this
behaviour.