Move module settings to profile config/install
In Open Social the votes of a user should be deleted when the user is
deleted. These votes are considered user data so we clean them up to
more easily comply with GDPR.

Open Social also provides some default configuration of where
like/dislike is enabled.

These settings.yml files shouldn't be installed when the profile is
installed. However, they should be picked up as overrides when the
respective module is enabled and tries to install that configuration
file. This happens when the Social Like module is enabled.

With the removal of features it's no longer possible to have existing
configuration files in a module's config/install folder as Drupal will
complain that the configuration already exists.