Add back jQuery UI styles in SocialBase
We used to remove some styles for jQuery UI. Probably because they broke
some things. However, we can not figure out what actually breaks, but we
do know that without these styles we can not properly support the layout
builder.

To remedy this the styles are added back in and the socialbase/modal
styles that create extra conflicts are removed from automatically
extending the jquery dialog library. To secure two cases where this
removal could possibly break things the library is manually added if the
dialog library is present.

This should preserve backwards behaviour as much as possibly while
allowing the layout builder to function without issues.