Add features removal post_update hook
When removing features we want to do a final features revert against a
known version of the configuration. This allows future configuration
updates to work from a known base.

Because features is already removed from each module (there's no
features.yml file anymore), it's not possible to simply call
features-revert. Luckily the `FeaturesManager::import` function can be
used instead to treat any module as a feature.

Because what's in the `config/install` folder can be different depending
on when this update hook is executed, we make a copy of the 8.0 version
of the folder and move it to `features_removal`. This ensures that even
if a platform is updated immediately to 8.4 (or another minor version)
the behaviour of this update hook remains the same.