Add static copy of 8.0 module config/install
This commit adds a static copy of the `config/install` folder in its 8.0 version.
This is used by `social_core_post_update_8800_features_revert` to perform a final
features revert to a known configuration state.

Any changes made to `config/install` before the 8.0 release should be copied to
this folder if the changed file was not under the `ignored` key of the
`<module>.features.yml` file.

This folder should not be changed after the 8.0 release is made. Instead use the
proper configuration workflow to update the site’s configuration.