Remove duplicated config and declare dependency
Social Post will fail without the Social Like module which was hidden by
our use of features and a duplicated config file. To ensure this doesn't
happen, a dependency is declared. A link is added to the issue that is
supposed to fix this properly.

Resolving this by declaring a dependency works fine because social_like
is a non-optional module in Open Social.

See Issue #3099988