Perform features revert in hook_update
Due to configuration changes in other modules we should not perform a
configuration change in post_update but we should do this in the update
hook. This also allows us to precisely control between which update
hooks the update is executed to ensure the features removal does not
introduce or remove configuration from other update hooks.