Let Drupal install optional modules
This removes all the custom logic around module installation. It inserts
a form that allows the selection of optional modules and inserts those
in the state value that Drupal uses to see which modules it should
install in the next step. This streamlines the installation process
which should result in improved performance and stability.

Additionally this removes all mention and usage of features from the
install process.