Move install profile code to Drupal namespace
Convention dictates that classes of the social profile should be under
`Drupal\social`. However, our custom psr-4 declaration in the
composer.json breaks this. The custom declaration is changed so things
resolve to the expected namespace.

The custom declaration shouldn't be needed, however it appears that
Drupal's automated resolution doesn't work yet if the profile is not yet
installed. This means that the Installer form class wouldn't be usable
early in the install process.