Move role creation into install profile config
This ensures that the roles are available when modules need them as
there's some configuration that relies on the configuration items.

The permission arrays are left empty as we already rely on all the
modules assigning default permissions to these roles. In the future we
may want to add a base set of permissions here and remove those from the
individual modules.