Remove PostInterface usage from activity_creator
The instanceof check was used to see if the module was loaded. However,
this can also be accomplished with a simple check against `NULL`. This
prevents the usage of the class from `post` and uses the storage
indirectly instead.

This small change allows the dependency on `social_post` to be removed
which is part of creating a circular dependency between
`activity_creator` and `activity_logger` that causes installation
problems.