Move group recipient field to optional config
This may break something in case the Social Group module is disabled
because I'm unsure where the field is used. However, it allows the
install to properly work because the activity_creator module is a
dependency of our Social entity modules and thus group will be installed
after.

The configuration does not belong in the social_group module because
it should be controlled by the activity_group module. Ideally this field
would be made independant of the entity type.