Move group recipient field back into config/install
The configuration is needed by the default view mode so it won't work in
config/optional. Instead we don't really care to what group type the
field references so we just remove the dependency on `social_group`
through the `open_group` type and depend only on the `group` module.

In the future this reference should be augmented further using a
`recipient resolver` or some plugin that uses a generic
`entity_reference` field without entity type.