Allow modules to declare themselves as optional features
This refactors the way the decision is made on what to show during
installation for users to enable as optional features. Modules can now
add a `modulename.social_feature.yml` file with `name`, `description`,
`weight` and `default` to indicate that they can be optionally enabled
during install time.

The installer will automatically pick up the module and add it to the
list of modules. The description is not currently used but may be shown
to users in the future to provide an explanation of the feature.

In case a module is not controlled by Open Social but should be
presented as a feature anyway then the module can be added to the
`social.feature_list.yml` file in the installation profile.