Set correct default landing page for social_user
The settings file had the value that was set in the social_user_install
hook. However, this value wasn't actually ever used because it didn't
validate against the configuration schema for the social_user module.
Now that the configuration was set with a configuration file, the config
was imported without validation. This caused the RouteSubscriber that
would previously default to `social_user.stream` if there was no value
to now use the incorrect value provided of the canonical user entity
route.

In addition to changing the default configuration the schema is updated
so that the configuration can be changed.