Don't depend on pathauto update when it's not installed
Drupal will skip all updates that are dependant on an update that is not
present. When pathauto is present we want all the update hooks of that
module to have been executed before running features revert. However, if
the module is not enabled adding the update hook dependency causes
problems. To solve this we can just ignore that update hook if pathauto
is not enabled.