Clean up demo content creation install steps
This commit removes the custom install finish step that was present. It
duplicated most of the things that Drupal core already does for us. The
only addition was indexing of the demo content.

This commit moves the indexing of the demo content into an install task
that's now dedicated to the demo content and that will only be visible
and run if the user actually selected demo content to be installed.

This reduces the amount of things we try to do ourselves and reduces the
chances for bugs to slip in.

As a bonus, due to the new timing of demo content installation, all
modules are already present so we can easily implement the install of
optional demo content by checking if the module that it depends on is
enabled.