Patch votingapi module to fix installation error
Our own installation process is now a bit more robust and sticks closer
to Drupal core. This does make errors during (un)installation in
underlying modules more prominent.

One such example is an issue in the votingapi module that was triggered
during the uninstall of the Social Demo module which managed to call
`votingapi_views_data_alter` that didn't handle the `vote_type` entity
not being available in the cache.