Move dependencies to .info.yml file for social_activity
The install hook of the Social Activity module installed some modules
that should probably be dependencies. The removed comment talked about
an issue however, this was probably features related which no longer
exists.

Looking at the new dependency graph it doesn't seem to cause issues to
just declare this as a proper dependency.