Mute content following messages for demo content
When demo content is created the social_follow_content module creates
messages that you're following the content. This can be quite disruptive
during installation. These messages are also not usually for the user
that is seeing them (i.e. the admin user).

To remedy this, when enabled, the social_demo module overrides the
Messenger service and doesn't show the message from the Social Follow
Content module.

The Social Demo module is supposed to be uninstalled directly after use
so this shouldn't cause any issues in production.