Mute all status messages in Social Demo
There are more modules that get triggered by demo content creation. It's
easier to just mute all messages (which we can assume is safe) rather
than try to add them all in the Social Demo Messenger class.

Only messages of type `status` are muted so that `error` and `warning`
messages do get through to the user.