Remove redundant entity load
The layout builder calls the preprocess hook with an unsaved entity.
This is fine because we already have the entity from our variables
array. However, it suddenly becomes apparent that using the entity you
have to load the new entity isn't needed. It's also problematic if the
entity is not yet persisted because then it can't be loaded.