Move bundle block theme suggestion before plugin
SocialBase added a theme suggestion based on the block bundle to allow
the AN hero block to receive a different template. However, this is
problematic because we want to style the block a certain way for the
layout builder.

This commit moves the most specific block plugin suggestion after the
generic one added by Open Social. This suggestion will vary by theme so
it shouldn't already exist unless someone was doing their own trickery
with theme suggestions.