Issue #3112214 by kokrull, dalfa, sjoerdvandervis: Hidden dependency on Flexible Group causes issues with feature revert and features removal update hook
Move the optional configuration to the Flexible Group module. This
ensures that features does not try to add the configuration during an
update. Features doesn't properly discriminate between required and
optional configuration.

This still breaks in the scenario where users have enabled the flexible
group module but not the landing page module. However, this scenario is
less likely.