Update view mode for new small teaser style.
This adds the update hooks and configuration changes for the layout
changes made to the Small Teaser style for all group types and
topic/event node types.

It relies on the Update Helper module to apply these changes. This
module will be added in another PR. Update hook dependencies will need
to be set to make that work.