Fix compatibility with Mink 1.8.0
Mink 1.8.0 no longer automatically starts a session when `getSession` is
called. This breaks some `beforeScenario` steps that rely on the
sessions (for example, testing JS support or resizing the window).

This commit simply checks whether the session is started before calling
any methods on the session, starting it if needed. This is only done in
`beforeScenario` calls that actually use the session.

For `FeatureContext` the beforeScenario calls are merged because I'm
unsure whether we can rely on call order.

Change in Mink: https://github.com/minkphp/Mink/pull/705