Issue #3111959 by chmez, Kingdutch, Ronaldtebrake, ribel: As a CM I want to customise filtering options in the custom content block (#1714)
* Implement dependency between fields

* Create a new plugin type for implementing dependencies between fields and add the first instance for topics

* Replace form fields for selecting a content type and filtering field by database fields

* Move filtering query to a method of new plugin type

* Add field to plugin definition for entity type ID

* Use the main theme instead of the admin theme for pages for creating and editing block

* Replace the bundle label field of plugin definition by the bundle ID field

* Add configuration of new fields to already enabled the module

* Fill in fields for plugin ID and plugin fields in existing blocks

* Restore the previous position of the database service between injected services to the builder service

* Issue #3111959 by chmez: Use entity base table instead of entity data table as the main table of the filtering query for has access to fields for sorting

* Issue #3111959 by chmez: Read a configuration from the code of update hook instead of reading configuration from YAML-files

* Update only blocks with an empty field for plugin ID

* Update a few blocks during on iteration of the update hook

* Add deleted phrases to the translations.php file

* Add comments to some parts of the code

* Rename topics variable to entities for the social-content-block template

* Apply default theming for the form actions section

* Fix bugs and fix dashboard usage of content list block

The block wouldn't render on layout builder pages because the correct
hook was not implemented.

Additionally, implementing that hook exposed some bugs. When no fields
were selected, the `empty` value of the field selector would not
actually be empty but would be interpreted as a special field. This is
fixed by checking for that field name, although this will break if a
field named `all` is every added. So it's probably better to just make
the field be actually empty.

Additionally a query alter would fail if there were no filters selected
because it attempted to do a join with an `IN` clause without and
values. This is not allowed in a SQL database.

* Delete empty fields before sending them to the plugin

* Add checking group content type

* Leave empty the field of the selected filter when all filters have been selected

* Add template file for content list block

This removes a custom template file for the content list block and
replaces it by a normal block template file.

It also begins the theming of the new small teaser look and feel for
topics/events/etc.

* Change text-muted color to #555

The text muted color of #777 often doesn't pass WCAG standards.
Darkening it slightly fixes this.

* Add content list block and small teaser styling

This changes the output of the small teaser and the content list block
for the new styles. The image is added as an element to be displayed.
This still requires a configuration update.

* Update view mode for new small teaser style.

This adds the update hooks and configuration changes for the layout
changes made to the Small Teaser style for all group types and
topic/event node types.

It relies on the Update Helper module to apply these changes. This
module will be added in another PR. Update hook dependencies will need
to be set to make that work.

* Add filter-fields on the create/edit block page automatically from the plugin definition

* Change position description field

* Add checking if the plugin management service exists before call it in the override service

* Improve update-files for completely hiding fields on the entity pages

* Add checking if the content of the block already exists

* Fix social_event CUD names

The names of the files was accidentally changed during a commit clean-up
but these should match the update hook name that they're executed in by
convention.

* Fix dependency on update helper update hook

Social Event update hook 8802 uses the update helper module but this is
not enabled until social_core_update_8805 so depending on
social_core_update_8802 could cause issues.

* Fix update-files for completely hiding fields on the entity pages

Co-authored-by: Roman Salo <frs.daddy@gmail.com>
Co-authored-by: Alexander Varwijk <alexander@getopensocial.com>
Co-authored-by: Serhii Myronets <serhii.myronets@lemberg.co.uk>