Add update and install hook to remove follow content button from topic small teassers
The follow content button is added after modules have been installed.
It's also the result of an optional module so it's possible that the
button doesn't exist. These factors don't allow the view display
configuration for the field to be set in configuration that is shipped
with our modules.

To work around this an modules_installed hook is implemented that makes
the change for us. An update hook is provided to fix platforms that
installed a version that had the new templates to render before this fix
was present.