Ensure sortBy function doesn't cause extra queries
The previous sortBy function already executed the query that was
prepared and then created another roundtrip to the database. However,
the things that were done can also be achieved by letting the database
create the join for us and saving a roundtrip. This has the advantage of
simplifying the code in cases where only a single extra table had to be
looked at.