Limit group content join to node relationships
There should be no duplicates in selected comments because a comment can
only belong to a single type of content and all joins are adding
distinctly typed relationships. This means that if `distinct` is needed
in a query we're masking a bug that is selecting the wrong comments.

In this case this occured because `group_content_field_data` contains
content related to a group through various relationship types. We are
interested in the `node` relationship type (all nodes belonging to a
group). However, we were also selecting the members of the group.
Entity ids between nodes and users can overlap so we were also selecting
any comments that were made by a member of the group as if those
comments were placed in the group itself.

To avoid this we must quantify which `type` of
`group_content_field_data` we wish to select. This must be done on the
join to that table instead of the later join to the voting/comment table
because otherwise the rows would still end up in the query.

It's chosen to select specifically for the `group_node` relationship
type instead of excluding the `group_membership` relationship type
because the latter may still select incorrect group content if a new
group content plugin is added. The former properly limits the join to
what we're interested in no matter what plugins are available. This
somewhat limits future flexibility, however this is already the case due
to the required specificity of the vote/comment join.