DS-7071 Implement event date based sorting
This commit allows per-content type sort options. The way this is done
is a bit tricky due to limitations in the sorting field that's added and
the way the form is build (if you come up with a better way, please
implement it!).

It then uses this to implement sorting based on event date that can be
used with an 'events after date' filter in a future commit to create an
'upcoming events' block.