Fix complementary region moving page content to the side
The socialblue_local_actions is rendered lazily, this can cause the
simple `if` check to  be false. This is not a problem in the normal
layout because the content is limited to 67% width and this region just
fills the remaining 1/3rd of the page.

However, in Sky this is problematic because the region is moved to the
left which causes the other content to be shifted to the right even if
there appears to be no actual content. To solve this we render the
contents of the region in the if statement. This is slightly slower but
is the only real way to find out if anything other than whitespace will
be rendered.