Issue #3119327 by Kingdutch: Apply styling for accept/decline buttons for group invitation
This removes the styling for the sky theme for the group join button.
The `transform: uppercase` was removed from Open Social quite some time
ago for a11y reasons and should not be re-introduced. Additionally the
`.btn-block` class exists for full width blocks. Using this per button
instead of forcing it from outside allows the dropdown for the
accept/join button that is desired.

The dropdown styling is also applied to the old flavour.