Issue #3096602 by Kingdutch: Fix admin toolbar and menu z-index fighting
We've been adding z-index fixes to the menu for the admin toolbar.
However, changing the z-index of the menu causes it to go under page
elements which wreaks havoc.

This commit looks at the underlying cause and stops trying to change the
menu but only changes the admin menu.

Unfortunately we do have to force the menu into relative positioning to
ensure the menu's aren't positioned on top of eachother.