Fix template suggestion for comment upload field group
The structure of the render array produced by the Field Group module has
changed which meant that our template didn't apply anymore causing
comment uploads to be inaccessible.

This commit changes the logic for adding the file suggestion, relying on
the consistency of the ID of the field group. `substr` is used because
a properly generated ID will get a `-*` number suffix when it's rendered
on a page multiple times, so an exact match may break in that case.