Force update graceful-fs to fix gulp@3
graceful-fs monkeypatches the Node fs module which no longer works in
newer versions of Node. It's not currently feasible to update gulp to a
newer version. However, the only change between graceful-fs 3 and 4
should be that it no longer monkeypatches `fs` which shouldn't matter
for the way gulp used it.

More information on this can be found here: https://stackoverflow.com/a/58394828/576060