Fix z-index for group teaser pop-over
The group teaser didn't have a z-index which caused it to be displayed
behind the secondary navigation bar. Giving it a z-index solves this
problem.

The chosen z-index is not related to anything else in Open Social so if
more issues are encountered putting this at a higher number is the way
to go. It should go above most page elements except a menu since those
shouldn't be open at the same time anyway as they both hide when
clicking on other elements on the page.