Remove teaser styling from group pop-over
The group pop-over very much looks like a teaser but applying the actual
styling from a teaser will break the pop-over. This is normally not a
problem because the module doesn't load the teaser styling. However,
when combined with Event Organisers this causes issues as that feature
does load the teaser styling.

The solution is to change the component itself instead of trying to
remove the teaser stylesheet as other future features may also need that
styling themselves.