Elaborate in pull request template
## Problem
It's not always clear for developers what the intention is of the headings in the pull request template. This causes information that can be useful to reviewers to be missed.

## Solution
Be more elaborate in the pull request template. Prompting the developer for the information that's expected and providing some examples.

A full How to test is included because it's sometimes difficult to determine the level of details required.

The issue link section is clarified to show that a Drupal.org issue is always needed for release management and links to other issue trackers are additions, not substitutions.

The screenshots section shows what a screenshot could be made of.

The release notes described better what they actually are.

The translations section now explains why it's there to guide developers in whether it's necessary or not.

## Issue tracker
GitHub only change not needed for the release notes so Drupal.org issue omitted.

## How to test
*For example*
- [ ] Ask a developer to make a pull request with the new template
- [ ] I expect it to be a subjective improvement over previous pull requests by that same developer