Enable helper modules when flexible group and / or social_book are enabled.
When social_group_flexible_group is enabled, we would like to enable the helper module social_flexible_group_featured, which is responsible for providing a featured view mode which is used when a flexible group is added as featured content on landingpages. The same goes for social_book, for which we enable the submodule social_book_featured.