Remove unrelated config from CUD
The intention of the Configuration Update Definition is to move the
like_and_dislike field into the correct position based on a fresh
install of Open Social. This should happen regardless of what
configuration is currently active.

The previous version of the CUD also tried to change some different
configuration that we don't want to touch. More importantly, it made
demands about this configuration that we don't care about. That could
result in this important update not being applied.

To resolve these problems we clear out the expected configuration so
that it should always apply. TWe also remove configuration changes not
related to like_and_dislike