Refactor multiple tests into single test
The piece of code under test is the same in all scenarios. We only test
different data inputs. A good way to do that would be to use a
dataprovider in PHPUnit but the set-up time for each test is pretty
high.

The data we test against isn't being mutated so there's no real reason
to have these tests as separate tests. Simply looping over the test data
and treating them as different assertions will not change the validity
of the test but saves about 80% of the test time :)