Add profile data to GraphQL API
This commit introduces the bulk of the profile fields in Open Social to
the GraphQL API. To support these fields some base types are introduced
in the social_graphql module. These types can be re-used by other parts
of the GraphQL schema that will be implemented in other modules. Those
types are not specific to social_user or social_profile.

Some more schema is written than is actually implemented in resolvers.
Most notably a start is made for user preferences. It provides a start
for how this could work but the various settings will need to be split
out to the modules that control them.

The different types of settings are split out to various types instead
of straight nested objects to allow additional modules to add fields as
needed (there may be extensions defining additional notifications for
example).

The locale settings are split out of the UserSettings object and
provided with a dedicated type since they may be useful on other
types than user. In multilingual applications these settings may also be
accessed more often than other user preferences so it helps to lift them
up a level.

The split between User and Profile is made at an "internal" vs
"external" level. The User type contains data and settings that control
the experience on the platform, as well as the bare minimum for certain
things to function (i.e. a display name and e-mail). The Profile type in
turn contains data that lets others know more about the user.