Issue #2947667 by ok_lyndsey, Kingdutch: Style skip to content so it is easily visible on each page (#2004)
The style is borrowed from Drupal.org. For the colors we've chosen to
use the navigation active link color as it's usually different from the
navbar color itself. Additionally it has a matching text color that
should be configured by platform users to something accessible. This has
the highest chance of resulting in a skip link that is styled to the
platform styles while being accessible.