Issue #3175829 by Kingdutch: It's unclear that dropdown menu items are expandable (#2027)
We look at the Bootstrap example and add the missing `aria-haspopup` and
`aria-expanded` attributes. The `aria-haspopup` attribute indicates to
assistive technology that there is a submenu present. The
`aria-expanded` attribute indicates the state of that submenu. The
`aria-expanded` attribute was already updated by the included JavaScript
but the initial `false` state was missing from the Twig rendered markup.

In addition we add the `role='button'` to indicate to screenreaders that
this is a menu button which makes it clearer that the button can be
pressed for more options.