Move event listener to body to prevent 'clickable' in voice over
When a click event handler exists on a parent of an element that is not
the body, OS X Voice Over (and possibly other screenreaders) will
announce any element selected as "clickable". This is quite disruptive
when going across teasers that contain non-clickable text.

Delegating the execution of the click handler to the body solves this
issue and no longer makes a string such as "Open Group" announce as
"Open Group, clickable".

In this case this is the right thing to do since the menu should only be
closable by clicking on the content on mobile, when the menu is open.
For screenreaders there are other ways to properly close the menu after
opening it.

Background: http://www.pauljadam.com/demos/voiceover-clickable.html