Issue #3184175 by Kingdutch: move from bower to npm
This replaces the packages that were pulled from bower to be now pulled
from NPM. Bower has been deprecated for a while and is quite unreliable
when used through asset-packagist. Packages from NPM should do better.

This was tested by committing the installed files to git, changing the
packages and then looking at the diffs. This revealed that the files
used in the library definitions are still shipped. The only detected
change is the renaming of `waves` to `node-waves` because the package
name is different on NPM.

An observed benefit of this change is that the source and test files are
no longer included (but the built and minified files are still present
in the NPM packages).