Implement standardised mutation payloads
This commit introduces our own Payload type to replace the GraphQL
module's Response type. The name `Payload` seems to be the expected
standard in JavaScript GraphQL clients.

Our own implementation specifies a Violation class to track errors
instead of requiring two inputs to the `addViolation` function. This
allows violations to be passed around. This flexibility is used in
extensions that implement mutations to separate the input validation
from the business logic of performing the mutation.

This commit also standardises on camelCase field naming (from
kebab_case) which follows the GraphQL best practices.

Finally this commit simplifies the `property_path` data producer usage
with `fromPath` calls. This is a change that's been recently merged into
the `graphql` module and should be in the next release.