Re-implement connection data fetching
The entity that is best used in Drupal to fetch the data for a
connection may not be the actual entity type that is returned on the
connection edges (or may more data is returned).

This commit reworks how the data loading happens and moves it out of the
data producer so we can use a slightly different format. The entity load
logic is now contained in the Entity Connection class itself. This means
we no longer need to pass on a lot of data from the thing that fetches
data to the resulting EntityConnection.

A new QueryHelper class is introduced that contains the actual data on
how to fetch a connection. This decouples the logic that implements
pagination from the best way to fetch the data based on the sorting.
This easily allows changing out the data fetching logic based on the
sorting field.