Fix pagination and ordering in `nodes` on connections
This edges result function did some post-processing for connection
results. This wasn't yet available on the `nodes` shortcut field which
caused results returned to clients for this field to be incorrect. The
logic is pulled out into a separate function and used by both.

We could've just called `getEdges` from `nodes` but that suggests
there's a required relationship, which is not necessarily true.