Add interface extension support to graphql-php
The GraphQL PHP library has a PR ready to implement support for
interfaces implementing other interfaces, which is supported by the
GraphQL spec. This is relied upon by some of the types in Open
Social.

We add the changes as a patch to the library so that we can use the
version that's shipped with drupal/graphql.