Require minimum 14.5.0 of webonyx/graphql-php
This commit removes the patch for interfaces implementing interfaces support
since the functionality has been merged into the module. The new minor version
is within the version constraints of `drupal/graphql` which causes it to be installed,
this then causes the patch to fail to apply.

We can safely remove the patch but must add our own requirement on a minimum
version to ensure that we have the new feature that we require. Using `>=` is fine here
because it allows `drupal/graphql` to supply the actual constraint and it allows us to 
update with that module to newer major versions, since the Drupal module contains most
of the surface that affects us in terms of compatibility.