Issue #3194731 by Kingdutch: [GraphQL] Improve nested pagination and sorting tests (#2185)
* Issue #3194731 by Kingdutch: [GraphQL] Improve nested pagination and sorting tests

* Ensure GraphQL tests fail loudly rather than silently

Tests could fail if a data producer caused an exception. The issue with
that is that an `assertEmpty` on the extensions field isn't very useful
because it provides no insight into _why_ the test failed. This means
that debugging a test failure in a CI environment can be very
cumbersome.

We add a little snippet of code that well re-throw any encountered
exceptions before performing our regular assertions. This ensures that a
developer can easily see the exception that caused the error.