Don't check interfaces for missing resolvers
Fields on interfaces are not required to have resolvers if all their
implementing types have resolvers for their fields, so it only makes
sense to check for resolvers on concrete types.

For orphaned resolvers we still want to check interfaces because
resolvers implemented for interfaces may be used to avoid repeating
resolvers for implementing types (for example `cursor` and `node` on any
`Edge` type).