Remove unimplemented GraphQL schema definitions
When we started we implemented quite a bit of schema that we haven't
wired up yet and that we may not need yet. This can cause confusion
among developers and API consumers.

The easiest way to solve this is to just remove these definitions, we
can always fish them out of git history.