Rename `currentUser` field to `viewer`
This is a shorter name and better matches other places of fields that
relate to information about the current actor. Other GraphQL APIs also
already commonly use this name for the field, most notably the GitHub
API.

Tests are adapted to use the built-in assertResults trait instead of
manually executing and matching the query. This requires adding some
cache metadata to match with. This cache metadata is not necessarily
what we want it to be yet but it provides us a place to change our
expectations and implement a fix.