Issue #3201182 by Kingdutch: [GraphQL] Introduce an Actor interface (#2258)
This commit introduces the `Actor` interface. An `Actor` indicates a
person or machine that has a representation on the platform and can
perform actions.

The `User` type is a real person implementation of an `Actor` indicating
a human user on the platform.

An example of a future `Actor` may be a `Bot` which could be used for
automated chat interactions or an `ExternalUser` which may be used to
post content from other platforms without having an account on the
platform itself.

As a result the `viewer` field is also turned into the `Actor` type to
prepare for clients not being real-users.

This commit introduces a `DecoratableTypeResolver` which should be used
for all type resolvers in Open Social's GraphQL API. This allows other
modules to extend the types that are returned for interfaces or unions
by decorating existing type resolvers.