Set .travis.yml open social package repo to test repo
composer checks the `composer.json` on the default repo branch to see if
it contains the package that we're looking for. This means that tests
fail for `drupal/social` since the main Open Social 10.1.x branch still
contains `goalgorilla/open_social`. This gives us a chicken/egg problem.

To work around this we use a different repo where we can already change
the package name and run the tests against that instead. This
demonstrates properly whether tests would fail or pass with the new
name.