Conflict with older goalgorilla/open_social versions
As the change to `drupal/social` is planned for introduction in 10.1.0
it's safe to replace any version between 10.0.0 and 11.0.0 according to
SemVer rules. The problem is that someone may be upgrading frmo an older
`goalgorilla/open_social` version which could cause `drupal/social` to
be installed next to it. This is undesirable as it could cause breakage
or hard to debug composer version conflicts. To solve this we make
`drupal/social` incompatible with older `goalgorilla/open_social`
versions.

The `>11.0.0` conflict is added because no versions of
`goalgorilla/open_social` should be added after 10.1.0 (but the replace
still allows libraries to depend on `goalgorilla/open_social`
successfully until 11.0.0)