Switch drupal_social to drupal/social
The master branch of drupal_social still depends on
`goalgorilla/open_social` at the time of this commit, which is needed
for other PRs to be able to complete testing. However, during the travis
set-up this causes composer to install both `drupal/social` and
`goalgorilla/open_social` in the update scenario, because we have no
good way to remove `goalgorilla/open_social` from `drupal_social`'s
composer.json file. Switching over to a different branch temporarily
fixes this issue.