Revert "Widen goalgorilla/open_social replace range"
This reverts commit ac13c56d87fbc583b48b8d9e9c3981e361f78584.

Widening the replace range caused composer to think that 10.1.x-dev can
replace 10.0.3, which the `replace` key says is true but it's not what
we want.

The widening was done to satisfy tests because earlier versions of
the repository were not available with the `drupal/social` name.
However, widening the replace range causes issue and is not the
solution. Instead we should've tagged older versions or fixed our
upgrade flow.

`drupal/social` is only ever compatible with the same version of
`goalgorilla/open_social`.