Modify GraphQL test to test Open Social comments
The `social_comment` module creates a comment type for nodes, so we use
the fields set-up by `social_comment` and the default node module to
test that our comment schema works.

We also stop testing as user 1 so that we must accurately implement in
tests which permissions are required (and can add further tests to test
the absense of these permissions).

For the bodyHtml of a comment in our fetch test we use a specific string
so that we can check that the output is properly processed. This reveals
that that's not the case. We're actually leaking Drupal's storage format
which isn't useful to consumers. This part of the test should probably
move out to its own test case so it's clearer when it fails.