Add TugboatQA configuration
This sets up a simple Open Social installation using a custom PHP image
based on open_social_docker but using the TugboatQA/php image as base
(the default php image is not supported).

The site is installed without optional modules so that it's easy to test
that modules only require the things they actually need or to build up
an environment as needed. Additionally we don't provide admin access
(although we can store a Tugboat configured admin password in LastPass
as back-up) to encourage testing with properly permissioned user.

As a compromise to allow modules to be enabled we provide this
permission to site managers. Addtionally we start out with demo content
so that there is a base to start manual testing from.

We use a custom composer.json file rather than trying to pull in
social_template so that we can control the tools we have available in
the preview environment more easily.